change layout polygons type

+++ smartroom_ml/utils/image_transform.py
@@ -341,14 +341,14 @@ def find_layout_polygons(layout_mask, blur_kernel=15, approx_strength=0.01):
         approx = cv2.approxPolyDP(contour, approx_strength * perimeter, True)
         approx = [{'point': [max(0, min(point[0][0] - 2, layout_mask.shape[1])),
                              max(0, min(point[0][1] - 2, layout_mask.shape[0]))], 'point_index': i,
-                   'point_classes': [plane_class]}
+                   'point_classes': [int(plane_class)]}
                   for i, point in enumerate(approx)]
-        layout_segments[plane_class] = approx
+        layout_segments[int(plane_class)] = approx
 
     layout_segments = fuse_points(layout_segments, max(layout_mask.shape) * 0.01)
     for layout_class, points in layout_segments.items():
-        layout_segments[layout_class] = [{'x': point['point'][0] / layout_mask.shape[1],
-                                          'y': point['point'][1] / layout_mask.shape[0],
+        layout_segments[int(layout_class)] = [{'x': float(point['point'][0] / layout_mask.shape[1]),
+                                          'y': float(point['point'][1] / layout_mask.shape[0]),
                                           'point_classes': point['point_classes']} for point in points]
     #     print(layout_segments[1])
     #     print(layout_segments[2])
@@ -374,7 +374,7 @@ def fuse_points(layout_segments, d):
                     continue
                 if dist2(points[i]['point'], points[j]['point']) < d2:
                     update_indexes.append([points[j]['point_index'], points[j]['point_classes'][0]])
-                    classes.append(points[j]['point_classes'][0])
+                    classes.append(int(points[j]['point_classes'][0]))
                     point['point'][0] += points[j]['point'][0]
                     point['point'][1] += points[j]['point'][1]
                     count += 1
@@ -382,8 +382,8 @@ def fuse_points(layout_segments, d):
             point['point'][0] /= count
             point['point'][1] /= count
             for i, layout_class in update_indexes:
-                layout_segments[layout_class][i]['point'] = point['point']
-                layout_segments[layout_class][i]['point_classes'] = list(set(classes))
+                layout_segments[int(layout_class)][i]['point'] = point['point']
+                layout_segments[int(layout_class)][i]['point_classes'] = list(set(classes))
     return layout_segments
